From retained backingmode to canvas image
Now the draw is made on a canvas image. That allows to show/hide the
palette
interface, to take a screenshot of the draw made using codea screenshot
functionality and in a future to add a save/load image functionality.
The drawback is the performance: from 56~ to 24~, less then half on iPad
2