MeshCanvas improved performances
Now MeshCanvas is Made of 2 different meshes. The first one is made of
just one rect with a texture that cover the whole screen. The second
mesh is used to memorize a single stroke, from start to end. On
brush:strokeEnd() the contents of this second mesh is drawn on the image
used as texture for the first mesh, and the second mesh is cleared. And
so on for each stroke.