New line implementation
segments of lines are now exactly fitting so to produce a better result
with both smoothed and not smoothed lines. If a line is made of
different segments an ID can be provide to the line draw call to notify
that and so the algorithm can easily apply correction to last segment.
Brushes has been modified to set a specific id for the main line,
letting all the other segments to be drawn without an id (so as single
segments).